pyproject.toml
[tool.poetry]
name = "burn-tokens"
version = "0.1.0"
description = "Burn Tokens Demo Web Application - A Flask-based demo web application. The name refers to LLM tokens consumed during AI-assisted development, not cryptocurrency tokens!"
authors = ["Generated by Poetry"]
readme = "README.md"
packages = [{include = "app.py"}]

[tool.poetry.requires-plugins]
poetry-plugin-export = ">=1.8"

[tool.poetry.dependencies]
python = "^3.12"
Flask = "3.0.0"
Flask-CORS = "6.0.0"
python-dotenv = "1.0.0"
requests = "2.32.4"
SQLAlchemy = "2.0.36"
Flask-SQLAlchemy = "3.1.1"
alembic = "1.14.0"
psycopg2-binary = "2.9.9"

[tool.poetry.group.dev.dependencies]
pytest = "7.4.3"
pytest-flask = "1.3.0"
flake8 = "6.1.0"
black = "24.3.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.flake8]
max-line-length = 88
extend-ignore = ["E203", "W503"]
exclude = [
    ".git",
    "__pycache__",
    ".venv",
    "venv",
    ".env",
    "migrations"
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = ["-v", "--tb=short"]
